Consolidated to use one Carousel Component. Outfit Carousel now shows an arrow only if there are enough cards in the carousel (>4).
