+++ client/src/components/RelatedItemsAndOutfit/UserOutfits/UserOutfit.jsx
@@ -1,20 +1,20 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
-import Carousel from './Carousel';
-import AddOutfitCard from './AddOutfitCard';
+import Carousel from '../Carousel';
+import AddOutfitCard from './AddOutFitCard';
 import OutfitListCard from './OutfitListCard';
 
-const UserOutfit = ({ currProd, prodsInfo, prodsStyles, prodsMeta }) => {
+const UserOutfit = ({ currProd }) => {
   const [currOutfits, setCurrOutfits] = useState([]);
   const addToCurrOutfits = (e) => {
     e.preventDefault();
     let alreadyAdded = false;
-    currOutfits.forEach((outfit) => {
-      if (outfit[0].id === currProd[0].id) {
-        alreadyAdded = true;
-      }
-    });
-    if (alreadyAdded) return;
+    // currOutfits.forEach((outfit) => {
+    //   if (outfit[0].id === currProd[0].id) {
+    //     alreadyAdded = true;
+    //   }
+    // });
+    // if (alreadyAdded) return;
     setCurrOutfits((oldList) => [...oldList, currProd]);
   };
   const removeOutfit = (idToRemove) => {
@@ -24,7 +24,7 @@ const UserOutfit = ({ currProd, prodsInfo, prodsStyles, prodsMeta }) => {
   return (
   <div className="outfit-list-container">
     Related Products Here:
-    <Carousel prodsInfo={prodsInfo} prodsStyles={prodsStyles} prodsMeta={prodsMeta}>
+    <Carousel>
       <AddOutfitCard
         addToCurrOutfits={addToCurrOutfits}
       />
@@ -42,9 +42,7 @@ const UserOutfit = ({ currProd, prodsInfo, prodsStyles, prodsMeta }) => {
 )}
 
 UserOutfit.propTypes = {
-  prodsInfo: PropTypes.instanceOf(Object).isRequired,
-  prodsStyles: PropTypes.instanceOf(Object).isRequired,
-  prodsMeta: PropTypes.instanceOf(Object).isRequired,
+  currProd: PropTypes.instanceOf(Array).isRequired,
 };
 
 export default UserOutfit;
